refactor: Add workersList param in engine struct
Add the `workerList` param in the `engine` struct to save the list of workers
passed in `Execute` method.

This enables the remove of the sort dependency that was needed to sort the
WorkerID keys in the workers map.

The WorkerID list was sorted to make the algorithm deterministic for test porpouses.

NOTE: According to the spec, "The iteration order over maps is not specified and is
      not guaranteed to be the sam from one iteration to the next."